pyproject.toml
# pyproject.toml
[tool.poetry]
name = "fitness-platform-backend"
version = "1.0.0"
description = "Fitness platform backend built with FastAPI and SQLAlchemy"
authors = ["Adonys Betancouer <adonys.cu2012@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
tenacity = "^8.2.3"  # Para retries
loguru = "^0.7.2"    # Logging profesional

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
ruff = "^0.1.6"
pre-commit = "^3.5.0"
factory-boy = "^3.3.0"
faker = "^20.1.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = "tests.*"
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E", "F", "W",  # pycodestyle
    "I",            # isort
    "B",            # flake8-bugbear
    "C4",           # flake8-comprehensions
    "UP",           # pyupgrade
]
ignore = ["E501", "B008"]
fixable = ["ALL"]
unfixable = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
    "slow: slow tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
